LF-8033: WW and command notification for VIT
This change was implemented to be able to know what is the WW
and command ID when the notify (SignalProcessor_notifyTrigger)
of these is called.
A new section of code was added to notify when a command is
detected, and as a parameter the WWId and commandID are sent.
Both the WW and the command notify are called as background to
avoid delays in the Voice_UI application.

Signed-off-by: Emmanuel Sanchez <emmanuel.sanchez@nxp.com>